+++ activity-log/internal/server/activity.go
@@ -13,10 +13,10 @@ type Activities struct {
 	activities []api.Activity
 }
 
-func (c *Activities) Insert(activity api.Activity) uint64 {
+func (c *Activities) Insert(activity api.Activity) int {
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	activity.ID = uint64(len(c.activities)) + 1
+	activity.ID = len(c.activities) + 1
 	c.activities = append(c.activities, activity)
 	log.Printf("Added %v", activity)
 	return activity.ID
@@ -24,11 +24,11 @@ func (c *Activities) Insert(activity api.Activity) uint64 {
 
 var ErrIDNotFound = errors.New("Id not found")
 
-func (c *Activities) Retrieve(id uint64) (api.Activity, error) {
+func (c *Activities) Retrieve(id int) (api.Activity, error) {
 	log.Printf("Getting %d", id)
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	if id > uint64(len(c.activities)) {
+	if id > len(c.activities) {
 		log.Printf("Id not found")
 		return api.Activity{}, ErrIDNotFound
 	}
